// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TransactionStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
  abandoned: 'abandoned'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]


export const TransactionType = {
  deposit: 'deposit',
  transfer: 'transfer'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]